refactor out RemoteFunctionSession
